Add option to not show date on content
Some content, like an about page or resume, should not be dated, so this
allows for that. Dates are removed from both the content page and the
list of posts.

To disable showing the date, set the `noshowdate` parameter to true in
the front matter of the content.

Originally authored by Jeremy Olexa <jolexa@jolexa.net> at
https://github.com/keichi/vienna/pull/28 on Dec 27 2015 and edited by
Ethan Madison <ethan@ethanmad.com> on May 10 2016.

Signed-off-by: Ethan Madison <ethan@ethanmad.com>